refactor: move findInactiveFields to core
