+++ packages/@tinacms/core/src/cms-forms/index.ts
@@ -18,3 +18,4 @@ limitations under the License.
 
 export * from './form-manager'
 export * from './form'
+export * from './findInactiveFields'
